refactor: Reward, BonusReward를 domain/rewards/ 폴더로 이동
