+++ src/domain/reward/BonusReward.js
