+++ src/domain/reward/Reward.js
